refactor: Move fetch files to services/fetches folder.
