+++ app/frontend/src/services/fetches/fetchDelete.js
